Fixed BarPlot extents calculation [doc-build]
