+++ holoviews/plotting/bokeh/chart.py
@@ -14,13 +14,12 @@ from ...selection import NoOpSelectionDisplay
 from ...core.data import Dataset
 from ...core.dimension import dimension_name
 from ...core.util import (
-    OrderedDict, max_range, basestring, dimension_sanitizer,
-    isfinite, range_pad, dimension_range)
-from ...element import Bars
+    OrderedDict, basestring, dimension_sanitizer, isfinite
+)
 from ...operation import interpolate_curve
 from ...util.transform import dim
-from ..mixins import AreaMixin, SpikesMixin
-from ..util import compute_sizes, get_min_distance, get_axis_padding
+from ..mixins import AreaMixin, BarsMixin, SpikesMixin
+from ..util import compute_sizes, get_min_distance
 from .element import ElementPlot, ColorbarPlot, LegendPlot
 from .styles import (expand_batched_style, line_properties, fill_properties,
                      mpl_to_bokeh, rgb2hex)
@@ -726,7 +725,7 @@ class SideSpikesPlot(SpikesPlot):
 
 
 
-class BarPlot(ColorbarPlot, LegendPlot):
+class BarPlot(BarsMixin, ColorbarPlot, LegendPlot):
     """
     BarPlot allows generating single- or multi-category
     bar Charts, by selecting which key dimensions are
@@ -762,51 +761,6 @@ class BarPlot(ColorbarPlot, LegendPlot):
     _y_range_type = Range1d
 
     selection_display = BokehOverlaySelectionDisplay()
-
-    def get_extents(self, element, ranges, range_type='combined'):
-        """
-        Make adjustments to plot extents by computing
-        stacked bar heights, adjusting the bar baseline
-        and forcing the x-axis to be categorical.
-        """
-        if self.batched:
-            overlay = self.current_frame
-            element = Bars(overlay.table(), kdims=element.kdims+overlay.kdims,
-                           vdims=element.vdims)
-            for kd in overlay.kdims:
-                ranges[kd.name]['combined'] = overlay.range(kd)
-
-        extents = super(BarPlot, self).get_extents(element, ranges, range_type)
-        xdim = element.kdims[0]
-        ydim = element.vdims[0]
-
-        # Compute stack heights
-        if self.stacked or self.stack_index:
-            ds = Dataset(element)
-            pos_range = ds.select(**{ydim.name: (0, None)}).aggregate(xdim, function=np.sum).range(ydim)
-            neg_range = ds.select(**{ydim.name: (None, 0)}).aggregate(xdim, function=np.sum).range(ydim)
-            y0, y1 = max_range([pos_range, neg_range])
-        else:
-            y0, y1 = ranges[ydim.name]['combined']
-
-        padding = 0 if self.overlaid else self.padding
-        _, ypad, _ = get_axis_padding(padding)
-        y0, y1 = range_pad(y0, y1, ypad, self.logy)
-
-        # Set y-baseline
-        if y0 < 0:
-            y1 = max([y1, 0])
-        elif self.logy:
-            y0 = (ydim.range[0] or (10**(np.log10(y1)-2)) if y1 else 0.01)
-        else:
-            y0 = 0
-
-        y0, y1 = dimension_range(y0, y1, self.ylim, (None, None))
-
-        # Ensure x-axis is picked up as categorical
-        x0 = xdim.pprint_value(extents[0])
-        x1 = xdim.pprint_value(extents[2])
-        return (x0, y0, x1, y1)
 
 
     def _get_factors(self, element, ranges):
